created a vec-common module
