+++ vec113/src/main/resources/vec113/vec_1.1.3-ext.xjb
@@ -20,12 +20,21 @@
         <jxb:property name="xmlId"/>
     </jxb:bindings>
 
-    <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Identification']]">
-        <inheritance:implements>com.foursoft.vecmodel.vec113.common.HasIdentification</inheritance:implements>
-    </jxb:bindings>
-    <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Description']]">
-        <inheritance:implements>com.foursoft.vecmodel.vec113.common.HasDescription</inheritance:implements>
-    </jxb:bindings>
+     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Identification']]">
+         <inheritance:implements>com.foursoft.vecmodel.common.HasIdentification</inheritance:implements>
+     </jxb:bindings>
+
+     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Description']]">
+         <inheritance:implements>com.foursoft.vecmodel.common.HasDescription&lt;com.foursoft.vecmodel.vec113.VecLanguageCode&gt;</inheritance:implements>
+     </jxb:bindings>
+
+     <jxb:bindings multiple="true" node="//xs:complexType[@name='AbstractLocalizedString']">
+         <inheritance:implements>com.foursoft.vecmodel.common.HasLocalizedString</inheritance:implements>
+     </jxb:bindings>
+
+     <jxb:bindings multiple="true" node="//xs:complexType[@name='LocalizedTypedString']">
+         <inheritance:implements>com.foursoft.vecmodel.common.HasLocalizedTypeString</inheritance:implements>
+     </jxb:bindings>
 
     <jxb:bindings node="//xs:complexType[@name='DocumentVersion']">
         <dg:delegate class="com.foursoft.vecmodel.vec113.delegate.VecDocumentVersionDelegate">
