+++ vec120/pom.xml
@@ -14,6 +14,11 @@
     <dependencies>
         <!--project-->
         <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>vec-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>com.foursoft.jaxb2</groupId>
             <artifactId>navigation-extender-runtime</artifactId>
         </dependency>
