+++ vec120/src/main/resources/vec120/vec_1.2.0-ext.xjb
@@ -21,19 +21,19 @@
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Identification']]">
-        <inheritance:implements>com.foursoft.vecmodel.vec120.common.HasIdentification</inheritance:implements>
+        <inheritance:implements>com.foursoft.vecmodel.common.HasIdentification</inheritance:implements>
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[.//xs:element[@name='Description']]">
-        <inheritance:implements>com.foursoft.vecmodel.vec120.common.HasDescription&lt;com.foursoft.vecmodel.vec120.VecLanguageCode&gt;</inheritance:implements>
+        <inheritance:implements>com.foursoft.vecmodel.common.HasDescription&lt;com.foursoft.vecmodel.vec120.VecLanguageCode&gt;</inheritance:implements>
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[@name='AbstractLocalizedString']">
-        <inheritance:implements>com.foursoft.vecmodel.vec120.common.HasLocalizedString</inheritance:implements>
+        <inheritance:implements>com.foursoft.vecmodel.common.HasLocalizedString</inheritance:implements>
     </jxb:bindings>
 
     <jxb:bindings multiple="true" node="//xs:complexType[@name='LocalizedTypedString']">
-        <inheritance:implements>com.foursoft.vecmodel.vec120.common.HasLocalizedTypeString</inheritance:implements>
+        <inheritance:implements>com.foursoft.vecmodel.common.HasLocalizedTypeString</inheritance:implements>
     </jxb:bindings>
 
     <jxb:bindings node="//xs:complexType[@name='DocumentVersion']">
